pyproject.toml
[project]
name = "dm-annotations"
version = "0.1.0"
description = "Discourse marker annotations and analysis model for academic Japanese text"
readme = "README.md"
authors = [
    { name = "Bor Hodošček", email = "dev@bor.space" }
]
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "anthropic>=0.64.0",
    "cupy-cuda12x>=13.4.1",
    "duckdb>=1.3.2",
    "ginza>=5.2.0",
    "ja-ginza>=5.2.0",
    # "ja-ginza-bert-large",
    "jaconv>=0.4.0",
    "matplotlib>=3.10.3",
    "netgraph>=4.13.2",
    "networkx>=3.4.2",
    "openai>=1.99.9",
    "openpyxl>=3.1.5",
    "orjson>=3.10.18",
    "pandoc>=2.4",
    "peewee",
    "polars[pyarrow]>=1.29.0",
    "prodigy",
    "pyarrow>=20.0.0",
    "pypandoc>=1.15",
    "pyrsistent>=0.20.0",
    "python-louvain>=0.16",
    "pyyaml>=6.0.2",
    "scikit-bio>=0.7.0",
    "spacy-transformers>=1.2.5",
    "sqlglot>=27.6.0",
    "statsmodels>=0.14.4",
    "tokenizers>=0.21.1",
    "torch>=2.7.0",
    "transformers>=4.17.0",
    "vegafusion>=2.0.2",
    "vl-convert-python>=1.8.0",
    "wtpsplit==2.1.6",
    "xlsxwriter>=3.2.3",
    "xxhash>=3.5.0",
]

[project.optional-dependencies]
marimo = [
    "altair",
    "marimo[recommended,lsp]==0.15.2", # FIXME: https://github.com/marimo-team/marimo/issues/6322
    "prince>=0.16.1",
    "scikit-bio>=0.7.0",
    "scikit-learn>=1.7.1",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/dm_annotations"]

[tool.uv.sources]
# ja-ginza-bert-large = { url = "https://github.com/megagonlabs/ginza/releases/download/v5.2.0/ja_ginza_bert_large-5.2.0b1-py3-none-any.whl" }
prodigy = { path = "wheels/prodigy-1.15.8-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl" }
peewee = { path = "wheels/peewee-3.16.3-py3-none-any.whl" }

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ty>=0.0.1a17",
    "types-networkx>=3.5.0.20250610",
    "types-openpyxl>=3.1.5.20250602",
    "types-tqdm>=4.67.0.20250516",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --doctest-modules"
testpaths = [
    "tests",
    "src",
]
doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"
disable_test_id_escaping_and_forfeit_all_rights_to_community_support = true

[tool.marimo.language_servers.pylsp]
enabled = true               # Enable/disable the Python language server
enable_mypy = true           # Type checking with mypy (enabled by default, if installed)
enable_ruff = true           # Linting with ruff (enabled by default, if installed)
enable_flake8 = false        # Linting with flake8
enable_pydocstyle = false    # Check docstring style
enable_pylint = false        # Linting with pylint
enable_pyflakes = false      # Syntax checking with pyflakes

[tool.marimo.diagnostics]
enabled = true               # Show diagnostics in the editor

[tool.marimo.runtime]
output_max_bytes = 10_000_000
